- DBContext is now working - Repositories little fixes

+++ trunk/SmartHome/SmartHome/BusinessEntities/BusinessHomeDevice/BusinessDoorLock.cs
@@ -10,7 +10,7 @@ namespace SmartHome.BusinessEntities.BusinessHomeDevice
         [OperationAttribute]        
         public static OperationMessage OpenDoor(this DoorLock doorLock)
         {
-            return OperationMessage.LogicWrite(doorLock.Id, LogicWriteValues.Set, doorLock.OpenTime);
+            return OperationMessage.LogicWrite((ushort)doorLock.Id, LogicWriteValues.Set, doorLock.OpenTime);
         }
     }
 }
